requirements.txt
streamlit
Pillow
piexif
streamlit-lottie